src/requirements.txt
# ---- Core API ----
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.2

# ---- Config & Models ----
pydantic==2.7.4
pydantic-settings==2.2.1

# ---- LLM ----
openai==1.30.0

# ---- Vector DB ----
qdrant-client==1.10.1

# ---- SQL ----
SQLAlchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0

# ---- Numeric ----
numpy==1.26.4

# ---- Documents ----
PyMuPDF==1.24.3
python-docx==1.1.0
rapidfuzz==3.9.4

# ---- NLP ----
spacy==3.7.5

sentence-transformers==2.6.1
transformers==4.38.2

--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.0+cpu


langchain-community


 en_core_web_lg==3.7.1
  camel-tools== 1.5.7
  camel_data -i all